adds new utils folder
