+++ .eslintrc.json
@@ -1,8 +1,8 @@
 {
   "env": {
     "browser": true,
-    "es6": true
-    // "node": true
+    "es6": true,
+    "node": true
   },
   "extends": [
     "airbnb-base",
